very experiments

+++ Runtime/Scripts/SceneExporter/ExporterAnimation.cs
@@ -901,7 +901,7 @@ namespace UnityGLTF
 
 					// Initialize data
 					// Bake and populate animation data
-					float[] times = null;
+					double[] times = null;
 
 					// arbitrary properties require the KHR_animation_pointer extension
 					bool sampledAnimationData = false;
@@ -1313,7 +1313,7 @@ namespace UnityGLTF
 			return curve;
 		}
 
-		private bool BakePropertyAnimation(PropertyCurve prop, float length, float bakingFramerate, float speedMultiplier, out float[] times, out object[] values)
+		private bool BakePropertyAnimation(PropertyCurve prop, float length, float bakingFramerate, float speedMultiplier, out double[] times, out object[] values)
 		{
 			times = null;
 			values = null;
@@ -1324,7 +1324,7 @@ namespace UnityGLTF
 			var nbSamples = Mathf.Max(1, Mathf.CeilToInt(length * bakingFramerate));
 			var deltaTime = length / nbSamples;
 
-			var _times = new List<float>(nbSamples * 2);
+			var _times = new List<double>(nbSamples * 2);
 			var _values = new List<object>(nbSamples * 2);
 
 			var curveCount = prop.curve.Count;
@@ -1540,14 +1540,14 @@ namespace UnityGLTF
 			return true;
 		}
 
-		internal static void RemoveUnneededKeyframes(ref float[] times, ref object[] values)
+		internal static void RemoveUnneededKeyframes(ref double[] times, ref object[] values)
 		{
 			if (times.Length <= 1)
 				return;
 
 			removeAnimationUnneededKeyframesMarker.Begin();
 
-			var t2 = new List<float>(times.Length);
+			var t2 = new List<double>(times.Length);
 			var v2 = new List<object>(values.Length);
 
 			var arraySize = values.Length / times.Length;
